refactor: strategy 패키지의 위치를 domain 패키지 내부로 변경
