+++ src/main/java/janggi/controller/mapper/ArrangementMapper.java
@@ -1,6 +1,6 @@
 package janggi.controller.mapper;
 
-import janggi.strategy.ArrangementOption;
+import janggi.domain.strategy.ArrangementOption;
 import janggi.view.label.ArrangementStrategyLabel;
 import java.util.EnumMap;
 import java.util.Map;
