+++ src/main/java/janggi/domain/strategy/BoardAssembler.java
@@ -1,4 +1,4 @@
-package janggi.strategy;
+package janggi.domain.strategy;
 
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.piece.Piece;
