+++ src/test/java/janggi/domain/board/BoardTest.java
@@ -7,8 +7,8 @@ import janggi.domain.Side;
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceType;
-import janggi.strategy.ArrangementStrategy;
-import janggi.strategy.BoardAssembler;
+import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.strategy.BoardAssembler;
 import janggi.support.TestArrangementStrategy;
 import janggi.support.TestPiece;
 import java.util.List;
