+++ src/test/java/janggi/strategy/BoardAssemblerTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.piece.Piece;
+import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.strategy.BoardAssembler;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
